+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 * Feature: Release notes added to NuGet packages
 * Fixed: Unchecked missing in `ValueObject` when claculating hash
+* Fixed: `NullReferenceException` thrown if `null` was stored
+  in `SingleValueObject` and `ToString()` was called
 
 ### New in 0.3 (released 2015-04-30)