+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 * Braking: `ValueObject` now uses public properties instead of both
   private and public fields
+* Braking: Aggregate IDs are no longer `string` but objects implementing
+  `IAggregateId`
 * Feature: Release notes added to NuGet packages
 * Fixed: Unchecked missing in `ValueObject` when claculating hash
 * Fixed: `NullReferenceException` thrown if `null` was stored